6 --- Autenticação com Token + Autorização/escola/requirements.txt
Django==2.2.9
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.1
pytz==2020.1
sqlparse==0.4.1
